HF_Deploy/requirements.txt
# ChatterboxTTS HuggingFace Spaces Requirements
# Optimized for HF Spaces environment with flexible versions

# Core ML and TTS - Essential (flexible versions for HF compatibility)
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.20.0
huggingface_hub>=0.15.0
safetensors>=0.3.0

# Audio processing - Required
soundfile>=0.12.0
librosa>=0.9.0
pydub>=0.25.0
audioread>=3.0.0

# ASR System - Intelligent ASR with fallback
openai-whisper>=20231117

# System monitoring and resource detection
psutil>=5.8.0
pynvml>=11.0.0

# Core scientific computing (flexible for HF environment)
numpy>=1.21.0
scipy>=1.7.0

# Text processing
regex>=2023.0.0
vaderSentiment>=3.3.0

# Web interface - Gradio (let HF manage version)
gradio>=4.0.0

# Progress and logging
tqdm>=4.60.0

# File handling
pathlib2>=2.3.0

# Configuration and utilities
python-dotenv>=1.0.0

# Optional utilities
requests>=2.25.0
packaging>=21.0

# Core ChatterboxTTS model dependencies
chatterbox-tts>=0.1.2
resemble-perth>=1.0.1
omegaconf>=2.3.0
einops>=0.6.0
diffusers>=0.21.0
tokenizers>=0.13.0
conformer>=0.3.0
s3tokenizer==0.2.0